Tidy tables.
